Renaming of the grid world environements
